Playgrounds/HW1-2_C3_ModelMNIST/requirements.txt
absl-py==0.7.1
aniso8601==6.0.0
astor==0.7.1
attrs==19.1.0
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
flask-restplus==0.12.1
gast==0.2.2
grpcio==1.20.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
MarkupSafe==1.1.1
mock==2.0.0
numpy==1.16.3
pbr==5.1.3
Pillow==6.0.0
protobuf==3.7.1
pyrsistent==0.14.11
pytz==2019.1
PyYAML==5.1
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.2
